+++ lib/install/template.rb
@@ -1,7 +1,7 @@
 # Install Webpacker
 copy_file "#{__dir__}/config/webpacker.yml", "config/webpacker.yml"
 
-say "Copying webpack core config and loaders"
+puts "Copying webpack core config"
 directory "#{__dir__}/config/webpack", "config/webpack"
 
 say "Copying .postcssrc.yml to app root directory"
@@ -17,29 +17,41 @@ say "Installing binstubs"
 run "bundle binstubs webpacker"
 
 say "Adding configurations"
-environment <<CONFIG, env: :development
-# Verifies that versions and hashed value of the package contents in the project's package.json
-config.webpacker.check_yarn_integrity = true
 
-CONFIG
-environment <<CONFIG, env: :production
+check_yarn_integrity_config = ->(value) { <<CONFIG }
 # Verifies that versions and hashed value of the package contents in the project's package.json
-config.webpacker.check_yarn_integrity = false
-
+  config.webpacker.check_yarn_integrity = #{value}
 CONFIG
+
+if Rails::VERSION::MAJOR >= 5
+  environment check_yarn_integrity_config.call("true"), env: :development
+  environment check_yarn_integrity_config.call("false"), env: :production
+else
+  inject_into_file "config/environments/development.rb", "\n  #{check_yarn_integrity_config.call("true")}", after: "Rails.application.configure do", verbose: false
+  inject_into_file "config/environments/production.rb", "\n  #{check_yarn_integrity_config.call("false")}", after: "Rails.application.configure do", verbose: false
+end
 
 if File.exists?(".gitignore")
   append_to_file ".gitignore", <<-EOS
 /public/packs
 /public/packs-test
 /node_modules
+yarn-debug.log*
+.yarn-integrity
 EOS
 end
 
 say "Installing all JavaScript dependencies"
-run "yarn add @rails/webpacker coffeescript@1.12.7"
+run "yarn add @rails/webpacker"
 
 say "Installing dev server for live reloading"
 run "yarn add --dev webpack-dev-server"
+
+if Rails::VERSION::MAJOR == 5 && Rails::VERSION::MINOR > 1
+  say "You need to allow webpack-dev-server host as allowed origin for connect-src.", :yellow
+  say "This can be done in Rails 5.2+ for development environment in the CSP initializer", :yellow
+  say "config/initializers/content_security_policy.rb with a snippet like this:", :yellow
+  say "p.connect_src :self, :https, \"http://localhost:3035\", \"ws://localhost:3035\" if Rails.env.development?", :yellow
+end
 
 say "Webpacker successfully installed 🎉 🍰", :green