+++ package/config.js
@@ -1,10 +1,24 @@
 const { resolve } = require('path')
 const { safeLoad } = require('js-yaml')
 const { readFileSync } = require('fs')
+const deepMerge = require('./utils/deep_merge')
+const { isArray } = require('./utils/helpers')
 
+const defaultFilePath = require.resolve('../lib/install/config/webpacker.yml')
 const filePath = resolve('config', 'webpacker.yml')
+
 const environment = process.env.NODE_ENV || 'development'
-const config = safeLoad(readFileSync(filePath), 'utf8')[environment]
+const defaultConfig = safeLoad(readFileSync(defaultFilePath), 'utf8')[environment]
+const appConfig = safeLoad(readFileSync(filePath), 'utf8')[environment]
+
+if (isArray(appConfig.extensions) && appConfig.extensions.length) {
+  delete defaultConfig.extensions
+} else {
+  /* eslint no-console: 0 */
+  console.warn('No extensions specified in webpacker.yml, using default extensions\n')
+}
+
+const config = deepMerge(defaultConfig, appConfig)
 
 const isBoolean = str => /^true/.test(str) || /^false/.test(str)
 
@@ -22,5 +36,8 @@ if (config.dev_server) {
     config.dev_server[key] = devServer(key)
   })
 }
+
+config.outputPath = resolve('public', config.public_output_path)
+config.publicPath = `/${config.public_output_path}/`.replace(/([^:]\/)\/+/g, '$1')
 
 module.exports = config