+++ package/rules/css.js
@@ -5,28 +5,35 @@ const { dev_server: devServer } = require('../config')
 const postcssConfigPath = path.resolve(process.cwd(), '.postcssrc.yml')
 const isProduction = process.env.NODE_ENV === 'production'
 const inDevServer = process.argv.find(v => v.includes('webpack-dev-server'))
-const extractCSS = !(inDevServer && (devServer && devServer.hmr))
+const isHMR = inDevServer && (devServer && devServer.hmr)
+const extractCSS = !(isHMR) || isProduction
+
+const styleLoader = {
+  loader: 'style-loader',
+  options: {
+    hmr: isHMR,
+    sourceMap: true
+  }
+}
 
 const extractOptions = {
-  fallback: 'style-loader',
+  fallback: styleLoader,
   use: [
-    { loader: 'css-loader', options: { minimize: isProduction } },
-    { loader: 'postcss-loader', options: { sourceMap: true, config: { path: postcssConfigPath } } },
-    'resolve-url-loader',
-    { loader: 'sass-loader', options: { sourceMap: true } }
+    { loader: 'css-loader', options: { minimize: isProduction, sourceMap: true, importLoaders: 2 } },
+    { loader: 'postcss-loader', options: { sourceMap: true, config: { path: postcssConfigPath } } }
   ]
 }
 
 // For production extract styles to a separate bundle
 const extractCSSLoader = {
-  test: /\.(scss|sass|css)$/i,
+  test: /\.(css)$/i,
   use: ExtractTextPlugin.extract(extractOptions)
 }
 
 // For hot-reloading use regular loaders
 const inlineCSSLoader = {
-  test: /\.(scss|sass|css)$/i,
-  use: ['style-loader'].concat(extractOptions.use)
+  test: /\.(css)$/i,
+  use: [styleLoader].concat(extractOptions.use)
 }
 
-module.exports = isProduction || extractCSS ? extractCSSLoader : inlineCSSLoader
+module.exports = extractCSS ? extractCSSLoader : inlineCSSLoader