+++ package/rules/file.js
@@ -1,6 +1,5 @@
 const { join } = require('path')
 const { source_path } = require('../config')
-const assetHost = require('../asset_host')
 
 module.exports = {
   test: /\.(jpg|jpeg|png|gif|tiff|ico|svg|eot|otf|ttf|woff|woff2)$/i,
@@ -8,8 +7,7 @@ module.exports = {
     loader: 'file-loader',
     options: {
       name: '[path][name]-[hash].[ext]',
-      context: join(source_path),
-      publicPath: assetHost.publicPathWithHost
+      context: join(source_path)
     }
   }]
 }